printer_qr_service/requirements.txt
asgiref==3.8.1
attrs==23.2.0
chardet==5.2.0
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
referencing==0.35.1
reportlab==4.2.2
reportlab-qrcode==0.1.6
rpds-py==0.19.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
